requirements.txt
# requirements.txt
# Core framework for the web UI
streamlit>=1.33.0,<2.0.0

# Google Generative AI SDK for interacting with Gemini models
google-generativeai>=0.5.0,<0.9.0

# Testing framework (only needed for development/CI)
# pytest>=7.0.0
